solucion error 500 en buscador roomie

+++ proyecto-limpio-spring/src/main/java/ar/edu/unlam/tallerweb1/repositorios/IRepositorioBuscador.java
@@ -12,7 +12,7 @@ public interface IRepositorioBuscador {
     Roomie ObtenerUnRoomie(String nombre);
     List<Usuario> ObtenerRoomies();
     Roomie buscarRoomie(String nombre);
-    List<Usuario> ObtenerRoomiesPorFiltro(Integer id);
+    List<Usuario> ObtenerRoomiesPorFiltro(Long id);
 
     void AgregarAlquiler(Vivienda vivienda);
     List<Vivienda> ObtenerAlquileres();
